canvas: fix the input registry issue on firefox
although I'm not a fan of how firefox fires tablet events, I
thought I'd fix it with a hack anyways.

firefox fires `pointerdown` events from a tablet pen in the
following order:
- buttons 1, type "touch";
- buttons 0, type "pen", which causes actionEnd to fire early
  and cause dots while drawing
any pointermove events afterwards are considered as buttons 1,
type "touch"

a fix for this would be to either refactor the pressing logic
to rely on the type of pointerEvent rather than its `buttons`
value, or to "lock" event processing to the pointerType of the
first pointerdown event. I'm not too keen on the former idea
for the time being, so this commit does the latter.