data: refactor clipping logic to fix improper clipping
the pushing logic in nln.forEach() only pushes points within the
sketch bounds, and those outside of it don't get properly clipped to
its edges. the lookbehind _intersections() in the first for loop don't
seem to help either. so instead, the pushing and intersection logic
should be separately handled by condition. additionally, this logic
could also be moved to that first for loop.

setting nln as the first point and starting the first for loop on the
second seems to be a redundant thing, why not just start the loop on
the first point?

and somehow, spawning of single-points by the clip gets patched by
the refactor. I'm assuming it's because of the strict conditioning.