css: remove center-alignment rule on `.settings-item > *`
the general rule to set all of a `.setting-item`'s children to
center-align its own children was removed, since this might cause
alignment issues for future UI changes to settings menus. i believe the
rule was actually supposed to align the item's children, so the rule's
set there instead just in case.