canvas: fix segmented-looking lines while drawing
a line that's being drawn will now be rerendered on the canvas as it's
being drawn -- before, it was drawn segment by segment, which made
vertices of the line look thick or gappy.

there's a catch with this commit's implementation though: putImageData
has pretty high latency and could cause lag if constantly called
(especially with lower sample rates), so for now, there'd be an option
to use that old rendering.

that lag is also why _animate() doesn't get the same treatment either.
but i think for visual's sake, this is better than nothing.