bring back removal of actions' `added-`/`removedPoints`
This reverts commit 4157a8c4ac57d7a2679e12d4641994107faf643d.

This branch had a couple big fixes that haven't been pushed in a while;
that commit specifically was a bit of a bottleneck to the stable push as
it was a bit of a breaking change.

I'm planning to bring it back once I dedicate more time to working on
the client.