data: return no intersections on boundary-laying lines
On _intersections(), if the given `line` has one of its endpoints right
on top of the `rect` bounds, it shouldn't be considered intersecting.
It doesn't make sense for it to be considered intersecting.

For _clip(), this should give _clip(), not _intersections(), the
responsibility of pushing points that are within canvas bounds into the
clip.