app: convert \s+ from pastes into single-spaces
This should turn a chain of spaces into a single space -- which is good
for preventing blank lines -- but this should also start supporting
non-0x20 spaces as well.

Someone on sketch stitches sketch data as part of their drawing
workflow. They usually use pass in \n's and they pass by the paste
check. These caused NaNs in <sketch-data>.lines down the road, so I
had to come in and deal with that.

There were two approaches to this, as far as I can tell:

1) Drop support for \s; only allow normal spaces.
2) Optimistically replace \s with normal spaces.

Took me a while to settle with that second one as a "Sure, I'll allow
that". Nothing wrong with making your data-stitching a bit more verbose.

Going for that first one feels like doing a middle-finger to that user,
especially since I don't have any proper changelog set up to
preemptively tell them about it.