app/tabs: replace pagehide listeners with beforeunload
The reason why 59975663 decided to replace the event to "pagehide" was,
if I remember correctly, because it was preferred by the Page Lifecycle
API over "beforeunload":
https://developer.chrome.com/docs/web-platform/page-lifecycle-api

Chrome had no issues with this change, but Firefox did: it tended to
not fully execute the entire event listener for some reason, so tab
closes were never properly registered ontp localStorage. Switching back
to "beforeunload" worked, though, so we're going for that. As far as
I've tried, the bfcache still works fine with it.

I might have to look into that a bit more someday.