api: timeout swaps that take a minute
These kinds of swaps tend to "swap" with another sketch, but don't
actually post the client's sketch onto the gallery. They usually take
around 60 seconds to get a response from garyc.me. Must be an anti-spam
feature.

Usually, sketches with vertical fills are prone to this bug. This can
be reproduced by filling the entire canvas with a 90deg fill and
swapping it right after.