app: pause saves when `.autosave` is off
.autosave tended to eat up saves when it's turned off, as it directly
influences .rememberTab and thus removes the save associated with the
current tab's ID. Opening more tabs with the setting still off would
have the same behavior happening with other IDs.

This change should ensure that .autosave only "pauses" the save process
instead of directly influencing .rememberTab. It wouldn't save or remove
entries when it's off. It wouldn't restore save entries either, but
rather it creates new ones in order to not make saves clash when
.autosave is turned back on.