zoom: fix 7406726's pointer capture issue
The "pointerout" issue pointed out by 7406726 (and re-introduced by
27e9aba) actually still exists. Testing was only done on desktop at
the time, but not on mobile, where the issue's *actually* prominent.
(I thought an issue like that would've extended to desktop.)

More specifically:

- The issue was invisible on hover-capable devices; a `pointerover`
  event would be fired right after the pointer capture release by
  <zoom-container>. The event would have a `.buttons` of `0`, so the
  client would act as if the pointer had just been let go.
- Mobile does *not* have hover capability. It doesn't have the benefit
  of sweeping the bug under the rug like that.

The revolutionary fix to that issue was to never make <canvas> pick up
panning movements in the first place.