app: permanently allow swapping of flipped sketches
This lock was originally meant for telling the user "Hey! Your canvas is
flipped. Don't keep on drawing while it's flipped!". It wasn't very
effective. It doesn't further help that if the user kept on drawing
without noticing the flip, they'd be forced to swap the thing flipped
around too.

In hindsight, I don't know why I thought that idea was any good.