app: use strings for `.backgroundType` instead of enums
This was bound to happen sooner or later; the numerical settings were
only made in the first place because I never found out about
`<paper-dropdown-menu>`'s `attr-for-selected` attribute. I thought I'd
be stuck with those forever...

There's a ton of number-to-string compatibility stuff sprinkled about,
so do take them away when it's due time.