add support for noz.rip's backup gallery
As of writing, garyc.me is down. It has been down since December 17th, and
no one's sure on when it's coming back. Right now, noz.rip is taking its
place, with a backup gallery having just started earlier today. This
backup will be the canonical gallery for the time being.

A new setting, `gallery`, is introduced in this commit. This replaces
the `gallerySource` setting, but there will be no fallback for it for
two reasons:

  - noz.rip is no longer a mirror to garyc.me's gallery.
  - noz.rip should be the default website to swap to while garyc.me is
    still down.

sketch-data's MAX_LENGTH is now writable too in order to reflect noz.rip's
128kB ink limit.