app: move majority of settings into a dedicated menu
The settings menu was getting a bit too cumbersome to navigate, so
this commit moves a big chunk of those settings into a separate dialog.

The CSS rules for the new menu are currently separate from the CSS rules
of the regular settings menu -- merging the two will have to be done in
a future commit.